mie

Module mie

Structs

CrossReference

SPARQL queries used for cross-references

DataStatistics

Statistics about the data

Mie

MIE: Metadata Interoperable Exchange Format This is a custom format to capture metadata about RDF datasets and their schemas, including shape expressions, example RDF data, SPARQL queries, and cross references. It is designed to be extensible and human-readable, facilitating sharing and analysis of RDF schema information. The main goal is to be used for MCP servers

RdfExample

RDF examples

SchemaInfo

Basic information about the schema and endpoint

ShapeExpression

Shape expressions defined in the schema

SparqlQueryExample

SPARQL queries as examples